Courses-Microservice-main/Courses/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.14.0
exceptiongroup==1.2.1
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==0.28.0
packaging==24.0
pillow==10.2.0
pydantic==2.7.0
pydantic_core==2.18.1
pytz==2024.1
requests==2.31.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
